requirements.txt
torch
torchvision
torchaudio
numpy
pandas
scipy
scikit_learn
matplotlib
tqdm
Pillow
opencv_python
opencv_contrib_python
kornia
regex
ftfy
wcwidth
colorama
captum
safetensors
adabelief_pytorch
git+https://github.com/openai/CLIP.git